requirements.txt
# Dataset
datasets==2.18.0
huggingface-hub==0.21.4

# Data processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
tqdm==4.66.2

# NLP
sentence-transformers==2.6.1
transformers==4.39.3
torch==2.2.1

# Backend
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
python-dotenv==1.0.1

# Evaluation
evaluate==0.4.1
rouge-score==0.1.2

# Notebook (optional but useful)
jupyter==1.0.0
ipykernel==6.29.3

# Utilities
python-multipart==0.0.9
httpx==0.27.0